+++ src/main/java/com/raoulvdberge/refinedstorage/tile/TileNode.java
@@ -102,7 +102,7 @@ public abstract class TileNode<N extends NetworkNode> extends TileBase implement
         INetworkNodeManager manager = API.instance().getNetworkNodeManager(world);
 
         NetworkNode node = (NetworkNode) manager.getNode(pos);
-        
+
         if (node == null) {
             manager.setNode(pos, node = createNode(world, pos));
             manager.markForSaving();
@@ -112,7 +112,16 @@ public abstract class TileNode<N extends NetworkNode> extends TileBase implement
             doLegacyCheck(node);
         }
 
-        return (N) node;
+        N castedNode;
+
+        try {
+            castedNode = (N) node;
+        } catch (ClassCastException e) {
+            manager.setNode(pos, castedNode = createNode(world, pos));
+            manager.markForSaving();
+        }
+
+        return castedNode;
     }
 
     private void doLegacyCheck(NetworkNode node) {